requirements.txt
PyYAML==6.0
pydantic==1.9.0
click~=8.0.4